requirements.txt
discord.py==2.6.2
aiohttp
python-dotenv
PyNaCl
pillow
transformers>=4.44.2
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
nltk